backend/requirements.txt
fastapi==0.99.1
uvicorn[standard]==0.18.3
python-dotenv==0.21.0
motor==2.5.1
solana==0.30.2
anchorpy==0.18.0
pydantic==1.10.12
solders==0.18.1 